Fix some type warnings

+++ .ino.cpp
@@ -2,7 +2,7 @@
 //This is a automatic generated file
 //Please do not modify this file
 //If you touch this file your change will be overwritten during the next build
-//This file has been generated on 2016-09-10 22:27:13
+//This file has been generated on 2016-09-10 22:49:28
 
 #include "Arduino.h"
 #include <adafruit_feather.h>
@@ -17,7 +17,7 @@
 #include "font.h"
 void setupSensor() ;
 void setup() ;
-void setChar(char c, char col, int *rgb) ;
+void setChar(char c, char col, uint16_t *rgb) ;
 void clearChar() ;
 void loop() ;
